requirements.txt
caldav==1.3.9
certifi==2024.2.2
charset-normalizer==3.3.2
icalendar==5.0.12
idna==3.7
lxml==5.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
recurring-ical-events==2.2.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tzlocal==5.2
urllib3==2.2.1
vobject==0.9.7
x-wr-timezone==0.0.7
